Refactor Assessment Utils
This puts all the table-related Utils into one file, and shared
utils that are date-related into a DateUtils file. This keeps
things a bit tidier. There’s probably scope for some more
tidiness, but my brain is melting now!
