+++ server/utils/assessments/utils.test.ts
@@ -2,35 +2,23 @@ import {
   acctAlertsFromAssessment,
   adjudicationsFromAssessment,
   allocationSummary,
-  assessmentLink,
   assessmentSections,
   assessmentsApproachingDue,
   assessmentsApproachingDueBadge,
-  awaitingAssessmentTableRows,
   caseNotesFromAssessment,
-  completedTableRows,
   confirmationPageMessage,
   confirmationPageResult,
-  daysSinceInfoRequest,
-  daysSinceReceived,
-  daysUntilDue,
-  formatDays,
-  formatDaysUntilDueWithWarning,
   formattedArrivalDate,
   getApplicationType,
   getPage,
   getReviewNavigationItems,
-  getStatus,
   getTaskResponsesAsSummaryListItems,
   groupAssessmements,
   rejectionRationaleFromAssessmentResponses,
-  requestedFurtherInformationTableRows,
   reviewApplicationSections,
 } from './utils'
 import { DateFormats } from '../dateUtils'
-import paths from '../../paths/assess'
 
-import * as personUtils from '../personUtils'
 import * as applicationUtils from '../applications/utils'
 
 import Assess from '../../form-pages/assess'
@@ -40,7 +28,6 @@ import {
   adjudicationFactory,
   applicationFactory,
   assessmentFactory,
-  clarificationNoteFactory,
   documentFactory,
   prisonCaseNotesFactory,
   userFactory,
@@ -55,7 +42,6 @@ const FirstPage = jest.fn()
 const SecondPage = jest.fn()
 
 jest.mock('../applications/utils')
-jest.mock('../personUtils')
 jest.mock('../reviewUtils')
 jest.mock('./documentUtils')
 jest.mock('../applications/arrivalDateFromApplication')
@@ -126,94 +112,6 @@ describe('utils', () => {
     })
   })
 
-  describe('daysSinceReceived', () => {
-    it('returns the difference in days since the assessment has been received', () => {
-      const assessment = assessmentFactory.createdXDaysAgo(10).build()
-
-      expect(daysSinceReceived(assessment)).toEqual(10)
-    })
-  })
-
-  describe('daysSinceInfoRequest', () => {
-    it('returns the difference in days since the assessment has been received', () => {
-      const today = new Date()
-
-      const date = new Date(today.getFullYear(), today.getMonth(), today.getDate() - 4)
-      const infoRequest = clarificationNoteFactory.build({ createdAt: DateFormats.dateObjToIsoDate(date) })
-      const assessment = assessmentFactory.build({
-        clarificationNotes: [clarificationNoteFactory.build(), infoRequest],
-      })
-
-      expect(daysSinceInfoRequest(assessment)).toEqual(4)
-    })
-
-    it('returns undefined if there are no info requests', () => {
-      const assessment = assessmentFactory.build({ clarificationNotes: [] })
-
-      expect(daysSinceInfoRequest(assessment)).toEqual(undefined)
-    })
-  })
-
-  describe('formatDays', () => {
-    it('returns the singular form if there is 1 day', () => {
-      expect(formatDays(1)).toEqual('1 Day')
-    })
-
-    it('returns the plural form if there is more than 1 day', () => {
-      expect(formatDays(22)).toEqual('22 Days')
-    })
-
-    it('returns N/A if the day in undefined', () => {
-      expect(formatDays(undefined)).toEqual('N/A')
-    })
-  })
-
-  describe('getStatus', () => {
-    it('returns Not started for an active assessment without data', () => {
-      const assessment = assessmentFactory.build({ status: 'active', data: undefined })
-
-      expect(getStatus(assessment)).toEqual('<strong class="govuk-tag govuk-tag--grey">Not started</strong>')
-    })
-
-    it('returns In Progress for an an active assessment with data', () => {
-      const assessment = assessmentFactory.build({ status: 'active' })
-
-      expect(getStatus(assessment)).toEqual('<strong class="govuk-tag govuk-tag--blue">In progress</strong>')
-    })
-
-    describe('completed assessments', () => {
-      it('returns "suitable" for an approved assessment assessment', () => {
-        const assessment = assessmentFactory.build({ status: 'completed', decision: 'accepted' })
-
-        expect(getStatus(assessment)).toEqual('<strong class="govuk-tag govuk-tag--green">Suitable</strong>')
-      })
-
-      it('returns "rejected" for an approved assessment assessment', () => {
-        const assessment = assessmentFactory.build({ status: 'completed', decision: 'rejected' })
-
-        expect(getStatus(assessment)).toEqual('<strong class="govuk-tag govuk-tag--red">Rejected</strong>')
-      })
-    })
-  })
-
-  describe('daysUntilDue', () => {
-    it('returns the days until the assessment is due', () => {
-      const assessment = assessmentFactory.createdXDaysAgo(2).build()
-
-      expect(daysUntilDue(assessment)).toEqual(7)
-    })
-  })
-
-  describe('formattedArrivalDate', () => {
-    it('returns the formatted arrival date from the application', () => {
-      const assessment = assessmentFactory.build()
-      ;(arrivalDateFromApplication as jest.Mock).mockReturnValue('2022-01-01')
-
-      expect(formattedArrivalDate(assessment)).toEqual('1 Jan 2022')
-      expect(arrivalDateFromApplication).toHaveBeenCalledWith(assessment.application)
-    })
-  })
-
   describe('getApplicationType', () => {
     it('returns standard when the application is not PIPE', () => {
       ;(applicationUtils.getApplicationType as jest.Mock).mockReturnValue('Standard')
@@ -232,76 +130,6 @@ describe('utils', () => {
       })
 
       expect(getApplicationType(assessment)).toEqual('PIPE')
-    })
-  })
-
-  describe('awaitingAssessmentTableRows', () => {
-    it('returns table rows for the assessments', () => {
-      const assessment = assessmentFactory.build()
-
-      ;(arrivalDateFromApplication as jest.Mock).mockReturnValue('2022-01-01')
-
-      const tierBadgeSpy = jest.spyOn(personUtils, 'tierBadge').mockReturnValue('TIER_BADGE')
-
-      expect(awaitingAssessmentTableRows([assessment])).toEqual([
-        [
-          { html: assessmentLink(assessment) },
-          { html: assessment.application.person.crn },
-          { html: 'TIER_BADGE' },
-          { text: formattedArrivalDate(assessment) },
-          { text: assessment.application.person.prisonName },
-          { html: formatDaysUntilDueWithWarning(assessment) },
-          { html: getStatus(assessment) },
-        ],
-      ])
-
-      expect(tierBadgeSpy).toHaveBeenCalledWith(assessment.application.risks.tier.value.level)
-    })
-  })
-
-  describe('requestedInformationTableRows', () => {
-    it('returns table rows for the assessments', () => {
-      const assessment = assessmentFactory.build({ clarificationNotes: clarificationNoteFactory.buildList(2) })
-
-      ;(arrivalDateFromApplication as jest.Mock).mockReturnValue('2022-01-01')
-
-      const tierBadgeSpy = jest.spyOn(personUtils, 'tierBadge').mockReturnValue('TIER_BADGE')
-
-      expect(requestedFurtherInformationTableRows([assessment])).toEqual([
-        [
-          { html: assessmentLink(assessment) },
-          { html: assessment.application.person.crn },
-          { html: 'TIER_BADGE' },
-          { text: formattedArrivalDate(assessment) },
-          { text: formatDays(daysSinceReceived(assessment)) },
-          { text: formatDays(daysSinceInfoRequest(assessment)) },
-          { html: `<strong class="govuk-tag govuk-tag--yellow">Info Request</strong>` },
-        ],
-      ])
-
-      expect(tierBadgeSpy).toHaveBeenCalledWith(assessment.application.risks.tier.value.level)
-    })
-  })
-
-  describe('completedTableRows', () => {
-    it('returns table rows for the assessments', () => {
-      const assessment = assessmentFactory.build()
-
-      ;(arrivalDateFromApplication as jest.Mock).mockReturnValue('2022-01-01')
-
-      const tierBadgeSpy = jest.spyOn(personUtils, 'tierBadge').mockReturnValue('TIER_BADGE')
-
-      expect(completedTableRows([assessment])).toEqual([
-        [
-          { html: assessmentLink(assessment) },
-          { html: assessment.application.person.crn },
-          { html: 'TIER_BADGE' },
-          { text: formattedArrivalDate(assessment) },
-          { html: getStatus(assessment) },
-        ],
-      ])
-
-      expect(tierBadgeSpy).toHaveBeenCalledWith(assessment.application.risks.tier.value.level)
     })
   })
 
@@ -333,48 +161,6 @@ describe('utils', () => {
       expect(assessmentsApproachingDueBadge(assessments)).toEqual(
         '<span id="notifications" class="moj-notification-badge">2<span class="govuk-visually-hidden"> assessments approaching due date</span></span>',
       )
-    })
-  })
-
-  describe('formatDaysUntilDueWithWarning', () => {
-    it('returns the number of days without a warning if the due date is not soon', () => {
-      const assessment = assessmentFactory.build({
-        createdAt: DateFormats.dateObjToIsoDate(new Date()),
-      })
-
-      expect(formatDaysUntilDueWithWarning(assessment)).toEqual('9 Days')
-    })
-
-    it('returns the number of days with a warning if the due date is soon', () => {
-      const assessment = assessmentFactory.createdXDaysAgo(8).build()
-
-      expect(formatDaysUntilDueWithWarning(assessment)).toEqual(
-        '<strong class="assessments--index__warning">1 Day<span class="govuk-visually-hidden"> (Approaching due date)</span></strong>',
-      )
-    })
-  })
-
-  describe('assessmentLink', () => {
-    const assessment = assessmentFactory.build({ id: '123', application: { person: { name: 'John Wayne' } } })
-
-    it('returns a link to an assessment', () => {
-      expect(assessmentLink(assessment)).toMatchStringIgnoringWhitespace(`
-        <a href="${paths.assessments.show({ id: '123' })}" data-cy-assessmentId="123">John Wayne</a>
-      `)
-    })
-
-    it('allows custom text to be specified', () => {
-      expect(assessmentLink(assessment, 'My Text')).toMatchStringIgnoringWhitespace(`
-        <a href="${paths.assessments.show({ id: '123' })}" data-cy-assessmentId="123">My Text</a>
-      `)
-    })
-
-    it('allows custom text and hidden text to be specified', () => {
-      expect(assessmentLink(assessment, 'My Text', 'and some hidden text')).toMatchStringIgnoringWhitespace(`
-        <a href="${paths.assessments.show({
-          id: '123',
-        })}" data-cy-assessmentId="123">My Text <span class="govuk-visually-hidden">and some hidden text</span></a>
-      `)
     })
   })
